Continue working on connection

+++ safesql-connection/src/main/java/org/safesql/connection/Update.java
@@ -2,6 +2,6 @@ package org.safesql.connection;
 
 public interface Update extends Result<Tuple> {
 
-    int execute();
+    long execute();
 
 }
